+++ src/api.js
@@ -29,7 +29,7 @@ const API = class {
     return fetch(`${this._endPoint}/${url}`, {method, body, headers})
       .then(checkStatus)
       .catch((err) => {
-        console.error(`fetch error: ${err}`);
+        // console.error(`fetch error: ${err}`);
         throw err;
       });
   }
@@ -60,7 +60,7 @@ const API = class {
   }
 
   updateTask(data) {
-    console.log(data);
+    // console.log(data);
     return this._load({
       url: `points/${data.id}`,
       method: Method.PUT,
