+++ src/trip-points/trip-point-constants.js
@@ -80,28 +80,4 @@ api.getDestinations()
 // .then(() => console.log('Destinations', Destinations));
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 export {AUTHORIZATION, END_POINT, typeTripPoint, Offers, Destinations};
